requirements.txt
requests==2.27.1
python-telegram-bot==13.13
pyTelegramBotAPI==4.5.1